+++ app/controllers/message/polls_controller.rb
@@ -2,9 +2,10 @@
 
 class Message::PollsController < ApplicationController
   before_action :poll_option
-  filter_access_to :all, attribute_check: true, model: PollOption
 
   def vote
+    authorize User, :logged_in?
+
     poll_option.poll_message.with_lock do
       poll_option.poll_message.poll_votes.where(user: current_user).destroy_all
       poll_option.poll_votes.find_or_create_by!(user: current_user)
